Revert "wifi: update the insmod function to use finit_module"
This reverts commit 02bd5eb704c528eaa61a0043404566fb88d17a90.

The new way of insmoding modules in N is giving us a tough time to deal with them.
Hence, revert to old ways to load wifi driver module and hence have wifi working.

Building wifi driver in kernel would have been a better solution, but since the wifi driver was written specifically to be worked with as a module, along with platform and oem specific changes, preventing us from switching to opensource drivers, and then other hardships like finding driver entry-point etc, we resort to this patch to have wifi working, period.